Create Main Baseball Code
All of the cleaned data from ESPN is here (both with and without pitching), we can also make a cleaned code without hitters and just pitchers as well. 

For some reason, Aaron Judges WAR is a little messed up and doesn't go to where it is supposed to be, so that is something we need to figure out.